feat: page 내 project 페이지 컴포넌트 views 로 이동
